refactoring

+++ Fullstack/my-unsplash/server/controllers/imageContoller.js
@@ -2,7 +2,7 @@ const Image = require("../models/ImageModel");
 
 // handle errors
 const handleErrors = (err) => {
-    console.log(err)
+    console.log(err);
     let errors = { label: "", url: "" };
     if (err.message === "Please enter a valid url") {
         errors.url = err.message;
@@ -30,30 +30,20 @@ module.exports.getImages = (req, res, next) => {
 
 module.exports.filterImages = (req, res, next) => {
     const query = req.params.query;
-    //   { $text: { $search: query } }
-    Image.find({ $text: { $search: query } })
-        .sort({ createdAt: -1 })
-        .exec(function (err, docs) {
-            if (!err) {
-                res.json({ images: docs });
-            } else {
-                console.log(err);
-                res.status(400).json({ errors: handleErrors(err) });
-            }
+
+    Image.find({ label: { $regex: query, $options: "i" } })
+        .then((result) => {
+            res.json({ images: result });
+        })
+        .catch((err) => {
+            console.log(err);
+            res.status(400).json({ error: err });
         });
-    // .then((result) => {
-    //     console.log(res)
-    //     res.json({ images: result });
-    // })
-    // .catch((err) => {
-    //     console.log(err);
-    //     res.status(400).json({ errors: handleErrors(err) });
-    // });
 };
 
 module.exports.saveImage = (req, res, next) => {
     const image = new Image(req.body);
-    console.log(req.body)
+    console.log(req.body);
     image
         .save()
         .then((result) => {
@@ -64,12 +54,18 @@ module.exports.saveImage = (req, res, next) => {
         });
 };
 module.exports.deleteImage = (req, res, next) => {
-    const id = req.params.id;
-    Image.findByIdAndDelete(id)
-        .then((result) => {
-            res.json({ result });
-        })
-        .catch((err) => {
-            res.status(400).json("Invalid Id");
-        });
+    const password = req.params.password;
+    console.log(password, process.env.PASSWORD);
+    if (password === process.env.PASSWORD) {
+        const id = req.params.id;
+        Image.findByIdAndDelete(id)
+            .then((result) => {
+                res.json({ result });
+            })
+            .catch((err) => {
+                res.status(400).json({ error: "Item not found" });
+            });
+    } else {
+        res.status(400).json({ error: "Invalid Password" });
+    }
 };
